requirements/local.txt
Django==1.10
django-ckeditor==5.1.0
django-taggit==0.21.0
django-watson==1.2.3
Pillow==3.3.1
